Administrator/requirements.txt
absl-py==1.4.0
arff2pandas @ git+https://github.com/RogerZhangsc/arff2pandas.git@6698a3f57a9c227c8f38deb827b53b80b8c6a231
backcall==0.2.0
boto3==1.23.10
botocore==1.26.10
cachetools==4.2.4
certifi==2024.6.2
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
fire==0.6.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.22.0
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.48.2
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
joblib==1.1.1
kiwisolver==1.3.1
liac-arff==2.5.0
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
openai==0.8.0
opencv-python==4.5.2.52
pandas==1.1.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.5
pycryptodome==3.20.0
Pygments==2.14.0
pymongo==4.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.5.2
sahi==0.11.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0
smmap==5.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
thop>=0.1.1.post2209072238
threadpoolctl==3.1.0
torch==1.9.1
torchaudio>=0.9.1
torchvision==0.10.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.18
utils==1.0.2
wcwidth==0.2.13
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
